predictor/requirements.txt
numpy
tqdm
typing-extensions
black
pypdf
transformers
peft
datasets
evaluate
google-cloud-aiplatform
google-cloud-storage
flask
gunicorn
bert_score